# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reisetagebuch API Draft",
        "description": "This is a draft version of the Reisetagebuch API documentation. The API is still under development and may change in future releases. Use at your own risk!",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.1.0-draft",
        "x-logo": {
            "url": "https://raw.githubusercontent.com/HerrLevin/reisetagebuch/refs/heads/main/public/assets/logo.webp"
        }
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local server"
        }
    ],
    "paths": {
        "/account": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Get account",
                "description": "Get account details",
                "operationId": "getAccount",
                "responses": {
                    "200": {
                        "description": "Account details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mustVerifyEmail": {
                                            "description": "Indicates if the user must verify their email address",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "traewellingConnected": {
                                            "description": "Indicates if the user has connected their Traewelling account",
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Account"
                ],
                "summary": "Delete account",
                "description": "Delete the authenticated account",
                "operationId": "deleteAccount",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "description": "Current password of the account",
                                        "type": "string",
                                        "example": "your-current-password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "405": {
                        "description": "Account deletion failed"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Account"
                ],
                "summary": "Update account",
                "description": "Update account details",
                "operationId": "updateAccount",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfileUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/account/socialite/traewelling": {
            "delete": {
                "tags": [
                    "Account"
                ],
                "summary": "Disconnect Traewelling",
                "description": "Disconnect Traewelling from account",
                "operationId": "disconnectTraewelling",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "405": {
                        "description": "Disconnection failed"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
        "/invites": {
            "get": {
                "tags": [
                    "Invites"
                ],
                "summary": "List invites",
                "description": "List invite codes for the authenticated user",
                "operationId": "listInvites",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/InviteDto"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Invites"
                ],
                "summary": "Create invite",
                "description": "Create a new invite code",
                "operationId": "createInvite",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreInviteCodeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/invites/{inviteCode}": {
            "delete": {
                "tags": [
                    "Invites"
                ],
                "summary": "Delete invite",
                "description": "Delete an invite code",
                "operationId": "deleteInvite",
                "parameters": [
                    {
                        "name": "inviteCode",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProfileUpdateRequest": {
                "title": "Profile Update Request",
                "description": "Request schema for updating user profile information",
                "required": [
                    "name",
                    "username",
                    "email"
                ],
                "properties": {
                    "name": {
                        "description": "The full name of the user",
                        "type": "string",
                        "maxLength": 255,
                        "example": "John Doe"
                    },
                    "username": {
                        "description": "The unique username for the user (lowercase, alphanumeric, dashes, and underscores only)",
                        "type": "string",
                        "maxLength": 30,
                        "example": "john_doe"
                    },
                    "email": {
                        "description": "The email address of the user (lowercase)",
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "example": "mail@example.com"
                    }
                },
                "type": "object"
            },
...
            "StoreInviteCodeRequest": {
                "required": [],
                "properties": {
                    "expires_at": {
                        "description": "The expiration date of the invite code",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "InviteDto": {
                "required": [
                    "id",
                    "createdAt",
                    "expiresAt",
                    "usedAt"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the invite code",
                        "type": "string"
                    },
                    "createdAt": {
                        "description": "The creation timestamp of the invite code",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "expiresAt": {
                        "description": "The expiration timestamp of the invite code",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "usedAt": {
                        "description": "The timestamp when the invite code was used",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
    "security": [
        {
            "passport": []
        }
    ]
}